Implement #29
Setting the flag `percent: true` on a `stacked: true` bar chart will create a dummy column for each series that is scaled to 100% and use that instead. The dummy column has the same name as the original column but with `_percent` appended. These dummy values are created __before__ any virtual columns are created so that tooltips can be created using these dummy values.