Add overlay #58
To highlight the shape we previously just updated the style of the layer and moved it to the front. That doesn't behave well for overlapping polygons. So we replicate what we do in SVG/Hex maps and add an overlay layer that is styled with a border and doesn't process pointer events.